Fix regressions from Tom Braun's code
The month_name(), day_name(), holiday_name(), and trad_day_name()
methods were returning undef, not '', in cases where no name could be
derived. This was broken in version 0.21.

The week_number() method was broken while making it return 0 for
Midyear's day and the Overlithe in version 0.21.

An author regression test was added to make sure this stuff stayed
fixed.