Rename sub t() to __t(), the hard way.
t() is imported into the name space by
DateTime::Fiction::JRRTolkien::Shire::Types, but I did not want it to
show up as a method, and I am nervous about magic as deep as
namespace::autoclean. So I loaded the types library with no imports, and
then hand-imported t() by assigning it to *__t.