Re-implement compare*() as static methods
The previous implementation assumed $dt1->compare( $dt2 ) was valid. And
this is in fact currently provided by the code, as a consequence of
overloading the space ship operator. But the only documented use of
compare() and compare_ignore_floating() is as a static method with two
arguments. So that is what the re-implementation does. Again, though,
the implementation relies on the corresponding DateTime methods using
the public interface, so I may end up re-thinking yet again.